ai-backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.127.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.7.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
sentry-sdk==2.48.0
shellingham==1.5.4
starlette==0.50.0
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
watchfiles==1.1.1
websockets==15.0.1

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Database drivers and ORM
asyncpg>=0.31.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# PDF Processing
pdfplumber
pymupdf>=1.24.0  # PDF page render to PNG for visual chunks
reportlab>=4.0.0  # PDF generation for exam papers
matplotlib>=3.7.0  # Math expression rendering (LaTeX → PNG via mathtext)

# Embeddings and Semantic Analysis
sentence-transformers  # Legacy - keeping for backward compatibility during migration
numpy

# LLM Integration
ollama
openai>=1.0.0  # Used for Gemini (OpenAI-compatible API)
json-repair>=0.58.0  # Fix malformed JSON from LLM responses

# Vector Database
qdrant-client

# Document Parsing (unstructured library)
unstructured[pdf]>=0.10.0
python-pptx>=0.6.0
python-docx>=1.0.0
pdfminer.six>=20221105
pillow>=10.0.0
python-magic-bin>=0.4.14; sys_platform == "win32"  # Windows only
python-magic>=0.4.27; sys_platform != "win32"  # macOS/Linux
